Added IntPrefPDUForwardingTable as interface for distinct distance vector.
Added ExtPrefPDUForwardingTable as IntPrefPDUForwardingTable that allows lookups for one QoS to match entries of other QoS (or types). As parameters, the id of BestEffort can be selected (for lookup without QoS) and an xml with the relation between QoSId and accepted entries must be passed.
Entries in the PDUForwardingTable does not need to be from valid QoSIds, but can be types.
The order of the list of valid types for each QoSId also decides the order of preference between valid types if more than one can be selected for an entry.
E.g. having QoSIds 1,2,3,4, we can define 100 as "fast" and 101 as slow. If QoS 1 and 2 are low delay QoS Cubes, there have 100 as valid type, while if 3 and 4 are medium delay QoS Cubes, their valid types are both 101 and 100 (in that order). Then, when lookup for QoS 1 and 2 we are going to accept only entries for type 100 or fast links, while for 3 and 4 we are going to accept all entries (100 and 101, fast and slow links), but use always the slow links if both are "best" entries.

This ForwardingTable allows to store only one entry for multiple QoS Cubes, something useful when for example all QoSCubes are mapped to the same RMTPort, only that using distinct queues.

PD.PrefDistanceVector has been changed in order to use any PDUForwardingTable extending IntPrefPDUForwardingTable