First commit, Full Exceptions and Rules forwarding policy.
String addresses must be hexstrings, the policy uses the converted value as address.
The policy works with neighbour id rather than neighbour address. Users can also define distinct neighbour groups in order to group neighbours useful to reach the same destinations.
The policy allows to configure Exceptions or Forwarding entries matching the first N bits of the address.
Exceptions can be of multiple forms:
- Unreachable > explicit "cannot reach destination"
- Common FWD table entry > list of ids that reach destination
- ANY -> can use any neighbour
- GROUP -> can use any neighbour of a given group
- ANY\Exceptions -> can use any neighbour except those of the given list
- GROUP\Exceptions -> can use any neighbour of a group, except of the given list

In addition to exceptions, this policy can be extended to try a forwarding rule in case of not having any exception matching the destination.
The default rule is "unreachable".
Rules are simple functions that from the destination address and information on the current node, returns a tuple (Group, first, total) that defines a range of how to reach a destination.
The values of that tuple are considered as follows:
- Group : 0 -> Consider all neighbours, Others -> Consider specific group
- First : First index of the range
- Total : Number of index valid
E.g. Having (1,3,2) as rule result and group 1 defined as [1,3,5,6,7,9],  would mean that we can use the ids at positions 3 and 4 of group 1, in this case neighbours with id 6 and 7.

The use of rules can be seen at example Routing/GRE/SmallDC, where the default configuration uses a routing to populate the exceptions, but ClosS configuration uses rules to provide forwarding in an scenario without routing.
A future improvement would be to add exceptions to the rules based routing, enabling then to forward packets using exceptions whenever rules are not valid.