Scheduling tutorial done.
Modified RA to solve an old problem with different QoS definition
between layers. Now, when checking if a N-1 flow is capable of
supporting an N flow, it considers then the QoSId of the N-1 flow (there
is a double check there, first ask forwarding for a next hop and then
search a flow for that next hop. Previously, that latter check was done
with the QoSs of the N DIF instead of that of the N-1, so if those where
different, the query wold fail) Also, when requesting for existing N
flows (for reusing a valid one instead of creating a new one), now it
first computes the required QoS as before the query was done for the
undefined QoS.