Updating SmallNetwork3 scenario, fixing DTCPState dealocation, EFCPPolicySet...
Examples:
Updated SmallNetwork3 scenario to work with the latest version.
Moved EFCP policy definition to the .xml into QoSCube section.
Renamed ipcProcess[x] to ipcProcess0[x] in router1 and router2.

DTCPState.cc
Commented section when I tried to cancel and delete retransmission timer
upon dealocatin DTCPState. My assumption is that since the timer is
scheduled it will be discarded by the simulation core.
To elaborate more, there is problem with retransmission timer, because
it is scheduled (and therefore owned) by the DTCP module, but we need it
in the DTCPState.

EFCPPolicySet.cc/h + QoSCube
Enhanced the output with names of all of the EFCP policies.

MockEFCPI.cc/h
Added destructor for the connectionID