refactor(fa): redo FA naming, remove some signal emission
Attributes being PascalCase is not done everywhere, and just looks ugly
when PascalCase should be used for types and classes. Changed to use
camelCase instead, which is used everywhere else in the code for
attributes and variables.

Also removed the use of some signals, as they were used wrong. Direct
method calls make for more readable code (not jumping between two or
more files to find the thread of execution), and probably saves some
execution speed.