feat(fa): use connection status with neighbour instead of self enroll
Only checking if this IPCP is enrolled in any DIF will cause no
additional management flows to be allocated. This will cause problems in
any case where several neighbouring IPCPs are reachable over any number
of underlying DIFs, and the simulation will crash. As of now, there are
zero examples that make use of several flows being allocated over
different "links", and as such this error has likely not been
encountered.

Connecting to several neighbours over the same segment is a relevant
use-case with the Ethernet shim DIF, so this needs to be fixed to allow
simulation with switches.

A side-effect is that some simulation scenarios will have more events
due to management flow allocation being required in every case.