feat(fa): ask for management flows when none exists for forwarding
This commit allows the Flow Allocator to go through the same N-1
management allocation procedure on forwarding of flows. This is
necessary because forwarding was handled differently and requires
additional state to be recorded from the N-1 IPCP (like what type of
flow is being requested). This does not work well for the Ethernet shim
IPCP, which cannot convey any more state than what Ethernet manages
(basically end-point addresses and DIF name).